backend/requirements.txt
Django==4.2.16
django-apscheduler==0.7.0
django-bootstrap-v5==1.0.11
django-bootstrap5==24.3
django-cors-headers==4.4.0
django-crontab==0.7.1

python-decouple==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
cryptography==2.8
requests==2.28.1
psycopg2==2.9.9
